chore: Ok finally figured out a way to deal with the tty raw/cooked mode stuff
So this is a total hack and more me getting backend junk working but now I can
send an rpc to change loglevel at runtime. As well as send keys in the tty to do
the same.

It isn't great but it works well enough for my use case that I can finally start
working on the actua logic rather than silly infra.

Now off to rpc land to start working on a cli and a way for a client to monitor
status of stuff inside the daemon.

Should also probably figure out how to unit test stuff in bevy.