chore: yeet is a glorified cp command now.
Not a great one at that but for mvp just want to replace rsync initially.

I'll have that then do periodic re-checks to find new/changed files the hard
way. This will be the default operating state for yeet if/when
epoll/inotify/ebpf fails.

As well as what darwin will use as its default. Aka it'll be the least worst
option if things go boom.