checkpoint: checking in the start of ssh connection and forwarding crap
Going to work on a cli for monitor/status to talk to the daemon first but this
is an OK start, its full of crap though. I need to build a Retry entity system
with the observer bevy stuff and events.

As well as likely refactor out some shared copy/paste between connection and
forwards similarly. But as is tradition past jerk mitch is leaving that to
future angy at past jerk mitch.