chore: github action for pre-release binaries off main
For now its just static linux x86_64, macos aarch64, and windows x86_64.

Latter is entirely untested/brained through. Just here to be sure I at
least have an idea of the blast radius of that eventually and to keep me
honest.

I can add darwin x86_64 if I need to though that platforms on life
support as of the latest macos release so not sure it matters. Linux
aarch64 is one to add soonish too but I don't use that often so not too
worried about it for now.