chore: setup git hooks in the flake itself
For now all that happens is:
- for pre-commit treefmt runs
- for pre-push nix flake check -L runs

I have nix flake check -L setup to run in github actions too so its a bit of a
wash there but in the future if others make changes I want to catch things in CI
too. So lets waste some computation methinks. I should be golden for my own
commits/pushes at least. Or try to be.